requirements.txt
asgiref
certifi
charset-normalizer
colorama
contourpy
crispy-bootstrap5
cycler
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
fonttools
fsspec
huggingface-hub
idna
Jinja2
kiwisolver
Markdown
MarkupSafe
matplotlib
mpmath
networkx
numpy
packaging
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
pydub
pyparsing
pyserial
python-dateutil
PyYAML
regex
requests
safetensors
setuptools
six
SpeechRecognition
sqlparse
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3
virtualenv